refactor: change name component
